ADPCM: Fix MSVC warning

+++ src/audio/codec/ADPCM.cpp
@@ -92,7 +92,7 @@ aalError CodecADPCM::setHeader(void * _header) {
 	padding = 0;
 	sample_i = 0xffffffff;
 	
-	size_t nb = 1 << shift;
+	size_t nb = size_t(1) << shift;
 	
 	predictor = new char[nb];
 	delta = new s16[nb];
